Enable and fix "The enum constant NAME should have a corresponding case label"
Omitting cases from switch blocks without raising a warning can cause
unexpected issues if a new enum value is added and switch blocks are
not updated to include it.

Since [1] the gerrit project includes jgit as a submodule, and has this
warning enabled, resulting in 100s of warnings in the console.

Enable the warning here, and fix all the occurrences by adding explicit
case statements for the missing values. Add the missing cases in the
same block as the 'default', so that there is no functional change.

In instances where the switch block only considered a single value,
replace the switch with an if-block.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/227897

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/dfs/DfsRefRename.java
@@ -63,13 +63,11 @@ final class DfsRefRename extends RefRename {
 
 		destination.setExpectedOldObjectId(ObjectId.zeroId());
 		destination.setNewObjectId(source.getRef().getObjectId());
-		switch (destination.update()) {
-		case NEW:
+		if (destination.update().equals(Result.NEW)) {
 			source.delete();
 			return Result.RENAMED;
-
-		default:
-			return destination.getResult();
 		}
+
+		return destination.getResult();
 	}
 }
